pyproject.toml
[project]
name = "travel-planner"
version = "0.1.0"
description = "AI Travel Planner Agent powered by LangGraph & Gemini"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "langchain",
    "langgraph",
    "langchain-google-genai",
    "serpapi",
    "streamlit",
    "pydantic",
    "python-dotenv",
    "rich",
    "requests>=2.32.5",
    "streamlit-js-eval>=1.0.0",
    "prompt-toolkit>=3.0.52",
    "googlemaps",
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.7",
]
